+++ src/app/components/pages/about-page/about-page.component.ts
@@ -1,5 +1,6 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { RouterLink } from '@angular/router';
+import { CONTACT_EMAIL, MAILTO_HREF } from '../../../constants';
 import { ProsePageComponent } from '../prose-page/prose-page.component';
 
 @Component({
@@ -7,4 +8,7 @@ import { ProsePageComponent } from '../prose-page/prose-page.component';
   imports: [ProsePageComponent, RouterLink],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class AboutPageComponent {}
+export class AboutPageComponent {
+  readonly contactEmail = CONTACT_EMAIL;
+  readonly mailtoHref = MAILTO_HREF;
+}