+++ src/app/components/pages/privacy-page/privacy-page.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { RouterLink } from '@angular/router';
-import { GITHUB_URL } from '../../../constants';
+import { CONTACT_EMAIL, GITHUB_URL, MAILTO_HREF } from '../../../constants';
 import { ProsePageComponent } from '../prose-page/prose-page.component';
 
 @Component({
@@ -10,4 +10,6 @@ import { ProsePageComponent } from '../prose-page/prose-page.component';
 })
 export class PrivacyPageComponent {
   readonly githubUrl = GITHUB_URL;
+  readonly contactEmail = CONTACT_EMAIL;
+  readonly mailtoHref = MAILTO_HREF;
 }